+++ packages/plugin-gallery/web/src/lib/convert-item-data.js
@@ -1,4 +1,5 @@
 import { normalizeUrl } from 'wix-rich-content-common';
+import { getAbsoluteUrl } from './baseUrlConverter';
 
 /**
  * convertItemData - converts the old gallery metadata format to the new metaData format
@@ -38,6 +39,13 @@ export const convertItemData = ({ items, anchorTarget, relValue }) =>
           rel: 'noopener',
           url: normalizeUrl(item.url || ''),
         };
+        const {
+          pathname,
+          thumbnail: { pathname: thumbPathname } = {},
+        } = convertedData.metaData.poster;
+        if (pathname && thumbPathname) {
+          convertedData.metaData.poster = getAbsoluteUrl(thumbPathname, 'image');
+        }
       }
 
       convertedData.metaData.alt = metadata.altText;